refactor: extract test fakes into testing modules
- Create FakeAuthenticationRepository in client/auth/data/testing
- Move GroceryRemoteDataSource interface and remote models from
  grocery/data/impl to grocery/data/public so they can be accessed
  by the testing module
- Create FakeGroceryRemoteDataSource in client/grocery/data/testing
- Remove inline private fakes from GroceryRepositoryImplTest
