+++ client/grocery/data/impl/src/commonTest/kotlin/com/plusmobileapps/chefmate/grocery/data/impl/GroceryRepositoryImplTest.kt
@@ -5,22 +5,19 @@ package com.plusmobileapps.chefmate.grocery.data.impl
 
 import app.cash.turbine.test
 import com.plusmobileapps.chefmate.auth.data.AuthState
-import com.plusmobileapps.chefmate.auth.data.AuthenticationRepository
 import com.plusmobileapps.chefmate.auth.data.ChefMateUser
-import com.plusmobileapps.chefmate.auth.data.SignUpResult
+import com.plusmobileapps.chefmate.auth.data.testing.FakeAuthenticationRepository
 import com.plusmobileapps.chefmate.database.Database
 import com.plusmobileapps.chefmate.database.testing.createTestDatabase
-import com.plusmobileapps.chefmate.grocery.data.impl.remote.GroceryRemoteDataSource
-import com.plusmobileapps.chefmate.grocery.data.impl.remote.RemoteGroceryItem
-import com.plusmobileapps.chefmate.grocery.data.impl.remote.RemoteGroceryList
+import com.plusmobileapps.chefmate.grocery.data.remote.RemoteGroceryItem
+import com.plusmobileapps.chefmate.grocery.data.remote.RemoteGroceryList
+import com.plusmobileapps.chefmate.grocery.data.testing.FakeGroceryRemoteDataSource
 import com.plusmobileapps.chefmate.util.testing.FakeDateTimeUtil
 import io.kotest.matchers.shouldBe
 import kotlin.test.Test
 import kotlin.uuid.ExperimentalUuidApi
 import kotlin.uuid.Uuid
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.runTest
 
@@ -28,7 +25,7 @@ class GroceryRepositoryImplTest {
 
     private val testDispatcher = UnconfinedTestDispatcher()
     private val db: Database = createTestDatabase()
-    private val fakeAuth = FakeAuthRepository()
+    private val fakeAuth = FakeAuthenticationRepository()
     private val fakeRemote = FakeGroceryRemoteDataSource()
     private val dateTimeUtil = FakeDateTimeUtil()
 
@@ -207,84 +204,4 @@ class GroceryRepositoryImplTest {
                 items.map { it.name }.toSet() shouldBe setOf("Milk", "Eggs")
             }
         }
-}
-
-private class FakeAuthRepository : AuthenticationRepository {
-    private val _state = MutableStateFlow<AuthState>(AuthState.Unauthenticated)
-    override val state: StateFlow<AuthState> = _state
-
-    fun setState(state: AuthState) {
-        _state.value = state
-    }
-
-    override suspend fun signInWithEmailAndPassword(email: String, password: String) =
-        Result.success(Unit)
-
-    override suspend fun signUpWithEmailAndPassword(email: String, password: String) =
-        Result.success(SignUpResult.Success as SignUpResult)
-
-    override suspend fun signOut() {
-        _state.value = AuthState.Unauthenticated
-    }
-
-    override suspend fun sendPasswordResetEmail(email: String) = Result.success(Unit)
-}
-
-private class FakeGroceryRemoteDataSource : GroceryRemoteDataSource {
-    val remoteLists = mutableMapOf<String, MutableList<RemoteGroceryList>>()
-    val remoteItems = mutableMapOf<String, MutableList<RemoteGroceryItem>>()
-    private var nextListId = 1
-
-    override suspend fun ensureDefaultList(ownerId: String): String {
-        val lists = remoteLists.getOrPut(ownerId) { mutableListOf() }
-        if (lists.isNotEmpty()) return lists.first().id!!
-        val newList =
-            RemoteGroceryList(
-                id = "remote-default-${nextListId++}",
-                name = "My Grocery List",
-                ownerId = ownerId,
-            )
-        lists.add(newList)
-        return newList.id!!
-    }
-
-    override suspend fun upsertGroceryItem(item: RemoteGroceryItem): RemoteGroceryItem {
-        val items = remoteItems.getOrPut(item.listId) { mutableListOf() }
-        val existing = items.indexOfFirst { it.id == item.id || it.clientId == item.clientId }
-        val result = item.copy(id = item.id ?: "remote-item-${Uuid.random()}")
-        if (existing >= 0) {
-            items[existing] = result
-        } else {
-            items.add(result)
-        }
-        return result
-    }
-
-    override suspend fun deleteGroceryItem(remoteId: String) {
-        remoteItems.values.forEach { it.removeAll { item -> item.id == remoteId } }
-    }
-
-    override suspend fun fetchAllGroceryItems(listId: String): List<RemoteGroceryItem> =
-        remoteItems[listId].orEmpty()
-
-    override suspend fun createGroceryList(list: RemoteGroceryList): RemoteGroceryList {
-        val result = list.copy(id = list.id ?: "remote-list-${nextListId++}")
-        val ownerLists = remoteLists.getOrPut(list.ownerId) { mutableListOf() }
-        ownerLists.add(result)
-        return result
-    }
-
-    override suspend fun fetchGroceryLists(ownerId: String): List<RemoteGroceryList> =
-        remoteLists[ownerId].orEmpty()
-
-    override suspend fun deleteGroceryList(remoteId: String) {
-        remoteLists.values.forEach { it.removeAll { list -> list.id == remoteId } }
-    }
-
-    override suspend fun updateGroceryList(list: RemoteGroceryList): RemoteGroceryList {
-        val ownerLists = remoteLists[list.ownerId] ?: return list
-        val index = ownerLists.indexOfFirst { it.id == list.id }
-        if (index >= 0) ownerLists[index] = list
-        return list
-    }
 }
