+++ client/grocery/data/public/build.gradle.kts
@@ -1,8 +1,14 @@
-plugins { alias(libs.plugins.kmpLibrary) }
+plugins {
+    alias(libs.plugins.kmpLibrary)
+    alias(libs.plugins.kotlinSerialization)
+}
 
 kotlin {
     sourceSets {
-        commonMain.dependencies { implementation(libs.kotlin.coroutines.core) }
+        commonMain.dependencies {
+            implementation(libs.kotlin.coroutines.core)
+            implementation(libs.kotlinx.serialization.json)
+        }
         commonTest.dependencies { implementation(libs.kotlin.test) }
     }
 }
