+++ client/grocery/data/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/grocery/data/remote/RemoteGroceryModels.kt
@@ -1,4 +1,4 @@
-package com.plusmobileapps.chefmate.grocery.data.impl.remote
+package com.plusmobileapps.chefmate.grocery.data.remote
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
